CMake, Conan with vscode integration
