+++ CMakeLists(conan1).txt
@@ -1,0 +1,23 @@
+cmake_minimum_required(VERSION 3.14)
+add_definitions("-std=c++17")
+
+project(VisualizadorProcesos)
+
+set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+
+# Incluir directorios
+include_directories(${PROJECT_SOURCE_DIR}/inc/)
+
+include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
+conan_basic_setup()
+
+# Agregar tus archivos fuente
+add_executable(${PROJECT_NAME} src/process_utils.cpp src/main.cpp)
+
+# Enlazar con las bibliotecas instaladas por Conan
+target_link_libraries(${PROJECT_NAME} PRIVATE ${CONAN_LIBS})
+
+# Habilitar C++17
+target_compile_features(VisualizadorProcesos PRIVATE cxx_std_17)
