+++ conanfile.txt
@@ -1,0 +1,7 @@
+[requires]
+nlohmann_json/3.11.3
+cpp-httplib/0.14.3
+
+[generators]
+CMakeDeps
+CMakeToolchain
